Refactor logging system to reduce redundancy and improve organization
- Merged overlapping functionality from logging.py and logger.py into core.py
- Eliminated duplicate implementations of formatters and trackers
- Centralized configuration management in LogManager
- Enhanced integration with consistent configuration inheritance
- Preserved all advanced features while providing a cleaner API
